+++ substrafl/remote/register/__init__.py
@@ -1,5 +1,7 @@
+from substrafl.remote.register.register import add_metric
 from substrafl.remote.register.register import register_algo
 
 __all__ = [
     "register_algo",
+    "add_metric",
 ]