+++ substrafl/remote/substratools_methods.py
@@ -172,6 +172,21 @@ class RemoteDataMethod:
             **self.method_parameters,
         )
 
+    def score(
+        self,
+        inputs: TypedDict,
+        outputs: TypedDict,
+        task_properties: TypedDict,
+    ) -> None:
+
+        datasamples = inputs[InputIdentifiers.datasamples]
+        predictions_path = inputs[InputIdentifiers.predictions]
+
+        method_to_call = self.instance.score
+        perf = method_to_call(datasamples=datasamples, predictions_path=predictions_path)
+
+        tools.save_performance(perf, outputs[OutputIdentifiers.performance])
+
     def load_trunk_model(self, path: str) -> Any:
         """Load the trunk model from disk
 
@@ -217,3 +232,4 @@ class RemoteDataMethod:
 
         tools.register(self.train)
         tools.register(self.predict)
+        tools.register(self.score)