+++ CHANGELOG.md
@@ -9,6 +9,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- The metric registration is simplified. The user can now directly write a score function within
+  their script, and directly register it by specifying the right dependencies and permissions.
+  The score function must have `(datasamples, predictions_path)` as signature. (#47)
+
+  Example of new metric registration:
+
+  ```py
+  metric_deps = Dependency(pypi_dependencies=["numpy==1.23.1"])
+  permissions_metric = Permissions(public=True)
+
+  def mse(datasamples, predictions_path):
+      y_true = datasamples["target"]
+      y_pred = np.load(predictions_path)
+      return np.mean((y_true - y_pred)**2)
+
+
+  metric_key = add_metric(
+      client=substra_client,
+      permissions=permissions_metric,
+      dependencies=metric_deps,
+      metric_function=mse,
+  )
+  ```
+
 - doc on the model loading page (#40)
 - The round 0 is now exposed.
   Possibility to evaluate centralized strategies before any training (FedAvg, NR, Scaffold).